requirements.txt
Django==4.2.7
pre-commit==3.5.0
psycopg2==2.9.9
virtualenv==20.24.6
psycopg-binary==3.1.13
django-allauth==  0.58.2
django-htmx==1.17.2
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-cov==4.1.0
